reorganize TypedCrossfilter code
